pyproject.toml
[project]
name = "pearlalgo-dev-ai-agents"
version = "0.1.0"
description = "Futures- and commodities-first agent scaffold for research, backtesting, and live-trading extensions."
authors = [{ name = "PearlAlgo", email = "dev@pearlalgo.ai" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2",
    "numpy>=2.0",
    "pydantic>=2.8",
    "pydantic-settings>=2.6",
    "pandas-ta>=0.4.70",
    "backtesting>=0.3.3",
    "python-dotenv>=1.0",
    "rich>=13.7",
    "ib-insync>=0.9.86",
    "scipy>=1.12",
    "click>=8.1",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "ccxt>=4.0.0",
    "polygon-api-client>=1.0.0",
    "vectorbt>=0.25.0",
    "streamlit>=1.28.0",
    "python-telegram-bot>=20.0",
    "discord.py>=2.3.0",
    "groq>=0.4.0",
    "litellm>=1.0.0",
    "PyYAML>=6.0",
    "loguru>=0.7.0",
    "websockets>=12.0",
    "websocket-client>=1.6.0",
    "aiohttp>=3.9.0",
    "scikit-learn>=1.3.0",
    "ta-lib>=0.4.0",
    "redis>=5.0.0",
]

[project.scripts]
pearlalgo = "pearlalgo.cli.main:cli"

[project.optional-dependencies]
backtrader = ["backtrader>=1.9.78.123"]
dev = ["pytest>=8.3", "ruff>=0.6"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
